19 feature upgrade to yargs (#20)
* refactor(package): updated depensort to use yargs package

Updated the package to use the yargs package to enable us to easily update/add features

BREAKING CHANGE: updated the entry point to use yargs

fixes #19

* chore(formatting): [skip ci] updating format to be consistent

---------
