+++ package.json
@@ -110,17 +110,19 @@
 		"release": "semantic-release -e ./.releaserc.json",
 		"prepack": "yarn run build"
 	},
-	"dependencies": {},
+	"dependencies": {
+		"yargs": "^17.7.2"
+	},
 	"devDependencies": {
 		"jest": "^29.3.1",
 		"husky": "^8.0.3",
 		"jsdoc": "^4.0.0",
 		"rimraf": "^3.0.2",
-		"dotenv": "^16.0.3",
 		"eslint": "^8.29.0",
+		"dotenv": "^16.0.3",
+		"taffydb": "^2.7.3",
 		"ts-node": "^10.9.1",
 		"ts-jest": "^29.0.5",
-		"taffydb": "^2.7.3",
 		"nodemon": "^2.0.20",
 		"prettier": "2.8.1",
 		"commitizen": "^4.2.6",
@@ -134,10 +136,10 @@
 		"@commitlint/cli": "^17.4.2",
 		"semantic-release": "^19.0.3",
 		"eslint-config-node": "^4.1.0",
-		"eslint-plugin-import": "^2.26.0",
 		"semantic-release-cli": "^5.4.4",
-		"@semantic-release/git": "^10.0.1",
+		"eslint-plugin-import": "^2.26.0",
 		"@semantic-release/npm": "^8.0.3",
+		"@semantic-release/git": "^10.0.1",
 		"eslint-config-prettier": "^8.6.0",
 		"eslint-plugin-prettier": "^4.2.1",
 		"jsdoc-plugin-typescript": "^2.2.0",
@@ -153,8 +155,8 @@
 		"@semantic-release/release-notes-generator": "^10.0.3"
 	},
 	"resolutions": {
-		"npm": "8.11.0",
 		"pug": "3.0.1",
+		"npm": "8.11.0",
 		"got": "11.8.5",
 		"json5": "1.0.2",
 		"marked": "4.0.10",
