+++ src/index.ts
@@ -1,23 +1,15 @@
-#!/bin/bash/node
-import File from "./File"
-import Sorter from "./Sorter"
+#!/usr/bin/env node
 
-const depensort = () => {
-	const packageFile = new File(`${process.cwd()}/package.json`)
-	const packageValues = packageFile.read()
-	if (packageValues.dependencies) {
-		packageValues.dependencies = new Sorter(packageValues.dependencies).sort()
-	}
-	if (packageValues.devDependencies) {
-		packageValues.devDependencies = new Sorter(packageValues.devDependencies).sort()
-	}
-	if (packageValues.resolutions) {
-		packageValues.resolutions = new Sorter(packageValues.resolutions).sort()
-	}
-	if (packageValues.overrides) {
-		packageValues.overrides = new Sorter(packageValues.overrides).sort()
-	}
-	packageFile.write(packageValues)
-}
+import yargs from "yargs/yargs"
+
+const depensort = () =>
+	yargs(process.argv.slice(2))
+		.scriptName("depensort")
+		.usage("Usage: $0 [command]")
+		.commandDir("./commands", {
+			extensions: ["ts", "js"],
+			exclude: /\.spec\.ts$/,
+		})
+		.strict().argv
 
 depensort()
